% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sacReBLEU-package.R
\docType{package}
\name{sacReBLEU-package}
\alias{sacReBLEU}
\alias{sacReBLEU-package}
\title{sacReBLEU: What the Package Does in One 'Title Case' Line}
\description{
One paragraph description of what the package does as one or more full sentences.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LazerLambda/sacReBLEU}
  \item Report bugs at \url{https://github.com/LazerLambda/sacReBLEU/issues}
}

}
\keyword{internal}
