% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sacReBLEU-package.R
\docType{package}
\name{sacRebleu}
\alias{sacRebleu}
\alias{sacRebleu-package}
\title{sacRebleu: An R package for calculating BLEU scores}
\description{
This package provides functions for calculating BLEU scores,
a common metric for evaluating machine translation models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LazerLambda/sacRebleu}
  \item Report bugs at \url{https://github.com/LazerLambda/sacRebleu/issues}
}

}
\author{
\strong{Maintainer}: Philipp Koch \email{PhillKoch@protonmail.com}

}
\keyword{internal}
