% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleu.R
\name{validate_arguments}
\alias{validate_arguments}
\title{Validate Arguemnts}
\usage{
validate_arguments(weights, smoothing, n)
}
\arguments{
\item{weights}{Weight vector for `bleu_corpus` and `bleu_sentence` functions}

\item{smoothing}{Smoothing method for `bleu_corpus` and `bleu_sentence` functions}

\item{n}{N-gram for `bleu_corpus` and `bleu_sentence` functions}
}
\value{
A list with the validated arguments (weights and smoothing)
}
\description{
Validate Arguemnts
}
