% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleu.R
\name{validate_references}
\alias{validate_references}
\title{Validate References}
\usage{
validate_references(references, target)
}
\arguments{
\item{references}{A list of reference sentences.}

\item{target}{A vector of target lengths.}
}
\value{
A boolean value indicating if the references are valid.
}
\description{
Validate References
}
