% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleu.R
\name{validate_arguments}
\alias{validate_arguments}
\title{Validate Arguments}
\usage{
validate_arguments(weights, smoothing, n)
}
\arguments{
\item{weights}{Weight vector for 'bleu_corpus_ids' and 'bleu_sentence_ids' functions}

\item{smoothing}{Smoothing method for 'bleu_corpus_ids' and 'bleu_sentence_ids' functions}

\item{n}{N-gram for 'bleu_corpus_ids' and 'bleu_sentence_ids' functions}
}
\value{
A list with the validated arguments (weights and smoothing)
}
\description{
Validate Arguments
}
